moved classes so that Silex\Provider can be used as a sub-tree split independently of eveything else
