+++ src/Silex/Provider/Locale/LocaleListener.php
@@ -9,13 +9,12 @@
  * file that was distributed with this source code.
  */
 
-namespace Silex\EventListener;
+namespace Silex\Provider\Locale;
 
 use Symfony\Component\HttpKernel\Event\GetResponseEvent;
 use Symfony\Component\HttpKernel\EventListener\LocaleListener as BaseLocaleListener;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\Routing\RequestContextAwareInterface;
-use Silex\Application;
 
 /**
  * Initializes the locale based on the current request.
@@ -26,7 +25,7 @@ class LocaleListener extends BaseLocaleListener
 {
     protected $app;
 
-    public function __construct(Application $app, RequestContextAwareInterface $router = null, RequestStack $requestStack = null)
+    public function __construct(\Pimple $app, RequestContextAwareInterface $router = null, RequestStack $requestStack = null)
     {
         parent::__construct($app['locale'], $router, $requestStack);
