+++ src/Silex/Provider/SessionServiceProvider.php
@@ -13,8 +13,8 @@ namespace Silex\Provider;
 
 use Silex\Api\ServiceProviderInterface;
 use Silex\Api\EventListenerProviderInterface;
-use Silex\EventListener\SessionListener;
-use Silex\EventListener\TestSessionListener;
+use Silex\Provider\Session\SessionListener;
+use Silex\Provider\Session\TestSessionListener;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\HttpFoundation\Session\Storage\Handler\NativeFileSessionHandler;
 use Symfony\Component\HttpFoundation\Session\Storage\NativeSessionStorage;
