+++ tests/Silex/Tests/Provider/ValidatorServiceProviderTest.php
@@ -56,7 +56,7 @@ class ValidatorServiceProviderTest extends \PHPUnit_Framework_TestCase
      */
     public function testConstraintValidatorFactory($app)
     {
-        $this->assertInstanceOf('Silex\ConstraintValidatorFactory', $app['validator.validator_factory']);
+        $this->assertInstanceOf('Silex\Provider\Validator\ConstraintValidatorFactory', $app['validator.validator_factory']);
 
         $validator = $app['validator.validator_factory']->getInstance(new Custom());
         $this->assertInstanceOf('Silex\Tests\Provider\ValidatorServiceProviderTest\Constraint\CustomValidator', $validator);
